Move unit tests for input functionality to test_inputs.py
